refac: create message module and separeted files
