+++ message/__init__.py
@@ -1,0 +1,3 @@
+__all__ = [
+    'default', 'contact', 'self_ref', 'bn_create', 'catcher_rec'
+]
