+++ message/contact.py
@@ -1,0 +1,4 @@
+KEY = 'お問い合わせ'
+START = 'お問い合わせありがとうございます。\n' \
+          '運営が対応致しますので、続けてお問い合わせ内容をご入力ください！\n' \
+          '「終了」と入力してお問い合わせが終わります！'
